chore: rename messages to wire
