--- securedrop-protocol/protocol-minimal/src/messages.rs
@@ -1,3 +1,0 @@
-//! Protocol messages
-pub mod core;
-pub mod setup;
