+++ securedrop-protocol/protocol-minimal/src/wire.rs
@@ -1,0 +1,3 @@
+//! Wire protocol request/response types.
+pub mod core;
+pub mod setup;
