+++ securedrop-protocol/protocol-minimal/src/wire/core.rs
