+++ securedrop-protocol/protocol-minimal/src/wire/setup.rs
