+++ securedrop-protocol/protocol-minimal/tests/setup.rs
@@ -7,10 +7,10 @@ use rand_core::{CryptoRng, RngCore, SeedableRng};
 use securedrop_protocol_minimal::LEN_DH_ITEM;
 use securedrop_protocol_minimal::api::{Api, JournalistApi};
 use securedrop_protocol_minimal::keys::FPFKeyPair;
-use securedrop_protocol_minimal::messages::setup::{
+use securedrop_protocol_minimal::sign::{FpfOnNewsroom, Signature};
+use securedrop_protocol_minimal::wire::setup::{
     JournalistEphemeralKeyRequest, JournalistSetupRequest,
 };
-use securedrop_protocol_minimal::sign::{FpfOnNewsroom, Signature};
 
 use securedrop_protocol_minimal::VerifyingKey;
 use securedrop_protocol_minimal::server::Server;
